requirements.main.txt
# ============================================
# Main Flask App + Nightly Scraper Dependencies
# ============================================

# Core Web Framework
Flask==3.0.0
Werkzeug==3.0.1

# AI/LLM Integration
httpx==0.27.0
openai==1.52.0

# Database
pymongo==4.6.1

# Environment & Configuration
python-dotenv==1.0.0
pytz==2023.3

# HTTP Client
requests==2.31.0
urllib3==2.1.0

# Google Sheets Integration
gspread==5.12.4
oauth2client==4.1.3
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0

# Security & Performance
cryptography==41.0.7
certifi==2024.2.2

# Scheduler for Nightly Jobs
schedule==1.2.0

# Web Scraping (for nightly scraper and indexer)
beautifulsoup4==4.12.3
lxml==5.1.0

# Image Processing & ML for indexer
numpy<2
Pillow==10.2.0
faiss-cpu==1.7.4
sentence-transformers==2.3.1
google-generativeai
